+++ src/Header.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import SearchInput from './search-input';
+import SearchInput from './SearchInput';
 
 import styled from 'styled-components';
